latest Joerg Sonnenberger patches and some fix to the spooler

+++ socket.c
@@ -28,7 +28,7 @@ int bind_to_unix(char *socket_name, int listen_queue, int chmod_socket, int abst
                 exit(1);
         }
         if (abstract_socket == 0) {
-        	if (unlink(socket_name) != 0) {
+        	if (unlink(socket_name) != 0 && errno != ENOENT) {
                 	perror("unlink()");
                 }
         }
